requirements.txt
aioboto3==13.3.0
aiobotocore==2.16.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aioitertools==0.12.0
aiosignal==1.3.1
altair==4.0.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.1
attrs==23.2.0
Automat==24.8.1
Babel==2.14.0
bcc==0.18.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
black==24.8.0
bleach==6.1.0
blinker==1.4
boto3==1.35.81
botocore==1.35.81
Brlapi==0.8.3
cached-property==1.5.2
cachetools==5.5.0
cattrs==1.10.0
certifi==2024.6.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
command-not-found==0.3
constantly==23.10.4
contourpy==1.2.0
cryptography==44.0.0
cssselect==1.2.0
cupshelpers==1.0
cvzone==1.6.1
cycler==0.12.1
Cython==0.29.28
dataclasses-json==0.6.7
dbus-python==1.2.18
debugpy==1.8.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docutils==0.17.1
email_validator==2.2.0
entrypoints==0.4
evdev==1.4.0
exceptiongroup==1.2.0
execnb==0.1.6
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.111.0
fastapi-cli==0.0.4
fastcore==1.7.8
fastjsonschema==2.19.1
filelock==3.16.1
fonttools==4.47.2
fqdn==1.5.1
freezegun==1.0.0
frozenlist==1.4.1
geoip2==2.9.0
gevent==21.8.0
ghapi==1.0.6
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-translate==3.19.0
google-search-results==2.4.2
googleapis-common-protos==1.66.0
googlesearch-python==1.2.5
greenlet==3.0.3
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==22.0.0
h11==0.14.0
hidpidaemon==18.4.6
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
hyperlink==21.0.0
hyperon==0.2.0
idna==3.7
importlib-metadata==4.6.4
incremental==24.7.2
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
isort==5.13.2
itemadapter==0.10.0
itemloaders==1.3.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
json5==0.9.14
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kernelstub==3.1.4
keyman-config==16.0.147
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.13
langchain-text-splitters==0.2.2
langgraph==0.1.4
langsmith==0.1.82
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libsass==0.20.1
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
maxminddb==2.0.3
mdurl==0.1.2
mercurial==6.1.1
mistune==3.0.2
more-itertools==8.10.0
mpld3==0.5.10
multidict==6.0.5
mypy==1.11.2
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbdev==2.3.31
nbformat==5.9.2
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
nose==1.3.7
notebook==7.0.7
notebook_shim==0.2.3
num2words==0.5.10
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.0
odoo==17.0+e.20240219
ofxparse==0.21
olefile==0.46
onboard==1.4.1
openai==0.28.0
openapi-schema-pydantic==1.2.4
opencv-python==4.9.0.80
orderedmultidict==1.0.1
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.3
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.5.0
ply==3.11
polib==1.1.1
pop-transition==1.1.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
propcache==0.2.0
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.29.2
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-keyring-linux-secretservice==0.1.0
proton-vpn-api-core==0.32.2
proton-vpn-gtk-app==4.4.4
proton-vpn-killswitch-network-manager==0.5.4
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.5.2
proton-vpn-network-manager-openvpn==0.1.0
psutil==5.9.0
psycopg2==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.7.4
pydantic-settings==2.6.1
pydantic_core==2.18.4
pydbus==0.6.0
pydeck==0.9.1
PyDispatcher==2.0.7
pydot==1.4.2
Pygments==2.18.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.1
PyMuPDF==1.25.1
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyOpenSSL==24.3.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyRFC3339==1.1
pyserial==3.5
pytest==8.3.3
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-json-logger==2.0.7
python-ldap==3.2.0
python-magic==0.4.24
python-multipart==0.0.9
python-stdnum==1.17
python-xlib==0.29
pytube==15.0.0
pytz==2022.1
pyusb==1.2.1.post1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.3.1
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.7.0
referencing==0.32.1
regex==2024.5.15
repolib==2.2.1
repoman==1.4.0
reportlab==3.6.8
requests==2.32.3
requests-cache==0.9.1
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rjsmin==1.1.0
roman==3.3
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.4
Scrapy==2.12.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semver==3.0.2
Send2Trash==1.8.2
sentry-sdk==1.4.3
service-identity==24.2.0
sessioninstaller==0.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.31
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.37.2
streamlit==1.18.1
supervisor==4.2.1
system76driver==20.4.106
systemd-python==234
tabulate==0.8.9
tenacity==8.4.2
terminado==0.18.0
tiktoken==0.7.0
tinycss2==1.2.1
tldextract==5.1.3
toml==0.10.2
tomli==2.0.1
toolz==1.0.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
Twisted==24.11.0
TwoCaptcha==0.0.1
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.21.0
validators==0.34.0
vboxapi==1.0
vobject==0.9.6.1
w3lib==2.2.1
wadllib==1.3.6
watchdog==5.0.3
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.0.2
widgetsnbextension==4.0.9
wrapt==1.17.0
wsproto==1.0.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
xmltodict==0.13.0
yarl==1.9.4
youtube-transcript-api==0.6.3
zeep==4.1.0
zipp==1.0.0
zope.event==4.4
zope.interface==5.4.0
